Fix React hydration error: extract Sidebar to client component, suppress body warnings
- Remove 'use client' from root layout (must be server component in App Router)
- Extract interactive Sidebar into components/Sidebar.tsx as proper 'use client' component
- Add suppressHydrationWarning to <body> to silence Grammarly extension attr injection
- Fix toLocaleDateString() calls to use explicit 'en-US' locale (prevents server/client mismatch)
